requirements.txt
flask
flask-session
langchain
langchain-community
chromadb
sentence-transformers
pymupdf
ebooklib
python-docx
ollama
plotly
fpdf2
umap-learn
scikit-learn
numpy
pandas
spacy
networkx
pyvis
rebel-large
